decremento e incremento disponibilita' libri

+++ SuorMaryWeb/src/it/beije/suormary/bookstore3/BookStoreUtility.java
@@ -211,14 +211,17 @@ public class BookStoreUtility {
     	   transaction.begin();
     	   try {		   
     		   Order order = entityManager.find(Order.class, orderId);
+    		   
    	       for(Book b : booksOrder) {
-    	    	    orderItem = new OrderItem();
+    	    	   orderItem = new OrderItem();
     	    	   orderItem.setBookId(b.getId());
     	    	   orderItem.setOrderId(order.getId());
     	    	   orderItem.setQuantity(b.getQuantity());
     	    	   orderItem.setPrice(b.getPrice() * b.getQuantity());
     	    	   entityManager.persist(orderItem);
     	    	   order.addOrderItem(orderItem);
+    	    	   Book book = entityManager.find(Book.class, b.getId());
+    	    	   book.setQuantity(book.getQuantity()-b.getQuantity());
     	       }
    	       double amount = 0;
    	       System.out.println(order.getItems().size());
@@ -285,6 +288,7 @@ public class BookStoreUtility {
        }
        public static void deleteOrder(int orderId) {
     	   EntityManager entityManager = JPAmanagerFactory.createEntityManager();
+    	   Book book = null;
     	   try {
     		   Order orderFound = entityManager.find(Order.class, orderId);
        	       EntityTransaction transaction = entityManager.getTransaction();
@@ -293,7 +297,11 @@ public class BookStoreUtility {
        	       query.setParameter("id", orderFound.getId());
        	       List<OrderItem> orderItems = query.getResultList();
        	       for(OrderItem orderItem : orderItems) {
-       	    	  entityManager.remove(orderItem);    	              	       
+       	    	 if(orderItem.getId() == orderItem.getId()) { 
+	   				  book = entityManager.find(Book.class, orderItem.getBookId());
+	   				  book.setQuantity(book.getQuantity() + orderItem.getQuantity());
+	       	    	  entityManager.remove(orderItem); 
+       	    	 }
        	       }
        	       entityManager.remove(orderFound);
        	       transaction.commit();
@@ -307,16 +315,22 @@ public class BookStoreUtility {
        public static void deleteOrderItem(String idStr) {
     	   EntityManager entityManager = JPAmanagerFactory.createEntityManager();
     	   int id = Integer.parseInt(idStr);
+    	     EntityTransaction transaction = entityManager.getTransaction();
+    		   transaction.begin();
+    	   Book book = null;
     	   try {
     		   OrderItem orderItem = entityManager.find(OrderItem.class, id);
     		   Order order = entityManager.find(Order.class, orderItem.getOrderId());
     		   for(OrderItem ord : order.getItems()) {
-    			   if(ord.getId() == orderItem.getId()) {
-    				   order.getItems().remove(ord);
+    			   if(ord.getId() == orderItem.getId()) { 
+    				  book = entityManager.find(Book.class, ord.getBookId());
+    				  book.setQuantity(book.getQuantity() + ord.getQuantity());
+    				  order.getItems().remove(ord);
     			   }
     		   }
-    		   EntityTransaction transaction = entityManager.getTransaction();
-    		   transaction.begin();
+    		 
+    		   
+	    	  
     		   entityManager.remove(orderItem);
     		   transaction.commit();
